requirements.txt
paho-mqtt>=2.0.0
flask>=3.0.0